[IMP] Integrate imps and fixes from @ahmadassaf git
The integration was made from: https://github.com/ahmadassaf/sublime-text-bootstrap
A Merge was not done as Ahmad removed the Windows support which I use

+++ setup.py
@@ -3,12 +3,13 @@
 import platform
 import sys
 import os
-import signal
 import subprocess
-import errno
-import shutil
-import urllib2
+import readline
+import json
+import codecs
+import copy
 
+from helpers import Helpers
 
 platform = platform.system()
 app_path = ''
@@ -18,9 +19,13 @@ config_path = ''
 
 
 def main():
+    print('Configuring SublimeText editor .....')
+    print('Checking your operating system .....')
+
     # OSX Setup:
     # - check and install via `brew cask`
     # - copy configuration to '/Library/Application Support/Sublime Text 3'
+
     if platform == 'Darwin':
         print('OSX detected')
         app_path = '/Applications/Sublime Text.app/Contents/SharedSupport/bin/subl'
@@ -30,10 +35,13 @@ def main():
             print('Sublime not installed...')
             install_osx(app_path)
         config(config_path)
+        configure_projects(config_path)
         sys.exit(0)
+
     # Linux Setup:
     # - check and install via `apt-get`
     # - copy configuration to '~/.config/sublime-text-3'
+
     if platform == 'Linux':
         print('Linux detected...')
         app_path = 'sublime'
@@ -42,7 +50,9 @@ def main():
             print('Sublime not installed...')
             install_linux(app_path)
         config(config_path)
+        configure_projects(config_path)
         sys.exit(0)
+
     # Windows Setup:
     # - check and notify user to install
     # - copy configuration to '%APPDATA%\Sublime Text 3'
@@ -54,27 +64,121 @@ def main():
             print('Sublime not installed...')
             install_windows(app_path)
         config(config_path)
+        configure_projects(config_path)
         sys.exit(0)
     else:
         raise Exception('Operating system not supported')
         sys.exit(1)
 
 
+# Automatically onfigure a list of projects into SublimeText
+def configure_projects(config_path):
+
+    def create_sublime_project(project_path, config_path):
+
+        # Define the path of the sublime project
+        sublime_session_path = config_path + "/Local/Session.sublime_session"
+        project_name = project_path.split('/')[-1]
+
+        # Create the .sublime-project data
+        sublime_project = {"folders": []}
+        sublime_project["folders"].append({"path": project_path})
+
+        sublime_workspace_path = project_path + \
+            "/" + project_name + ".sublime-workspace"
+        sublime_project_path = project_path + "/" + project_name + ".sublime-project"
+
+        # Create the .sublime-workspace which is just a default empty JSON
+        print("--> creating sublime project in: " + project_path)
+
+        with open(sublime_project_path, 'w') as project:
+            project.write(json.dumps(sublime_project, ensure_ascii=False))
+        with open(sublime_workspace_path, 'w') as workspace:
+            workspace.write(json.dumps({}, ensure_ascii=False))
+
+        # Now we need to make sure that we add this to the
+        # Session.sublime_session
+        if os.path.exists(sublime_session_path):
+            # We need just to append the new project names now
+            with codecs.open(sublime_session_path, 'rU', 'utf-8') as session_file:
+                sublime_session = copy.deepcopy(json.load(session_file))
+                sublime_session["workspaces"][
+                    "recent_workspaces"].append(sublime_workspace_path)
+                sublime_session["workspaces"]["recent_workspaces"] = Helpers.uniquify(
+                    sublime_session["workspaces"]["recent_workspaces"])
+                with codecs.open(sublime_session_path, 'w', encoding="utf-8") as session:
+                    session.write(json.dumps(
+                        sublime_session, ensure_ascii=False))
+        else:
+            # We need to create the file and fill it up at least for the first
+            # time
+            sublime_session = {"workspaces": {"recent_workspaces": []}}
+            sublime_session["workspaces"][
+                "recent_workspaces"].append(sublime_workspace_path)
+            sublime_session["workspaces"]["recent_workspaces"] = Helpers.uniquify(
+                sublime_session["workspaces"]["recent_workspaces"])
+            with open(sublime_session_path, 'w') as session:
+                session.write(json.dumps(sublime_session, ensure_ascii=False))
+
+    configure = raw_input(
+        "Would you like to setup automatically a list of projects into SublimeText ? [y/n] : ")
+
+    if configure == "y":
+        choice = raw_input(
+            "Would you like to setup multiple projects or single project ? [m(multiple)/s(single)] : ")
+        if choice == "m":
+            while True:
+                readline.set_completer_delims(' \t\n;')
+                readline.parse_and_bind("tab: complete")
+                readline.set_completer(Helpers.complete)
+                parent_folder = raw_input(
+                    "Please enter destination of the parent folder. Note that all subfolders will be added as project (type 'exit' anytime to quit): ")
+                if parent_folder.strip() == 'exit':
+                    break
+                if os.path.exists(parent_folder) and os.path.isdir(parent_folder):
+                    # List all folders immediately under this folder:
+                    dirs = (next(os.walk(parent_folder))[1])
+                    for dir in dirs:
+                        if dir.startswith("."):
+                            continue
+                        create_sublime_project(
+                            parent_folder + "/" + dir, config_path)
+                else:
+                    print("this does not seem to be a valid path :(")
+        elif choice == "s":
+            while True:
+                readline.set_completer_delims(' \t\n;')
+                readline.parse_and_bind("tab: complete")
+                readline.set_completer(Helpers.complete)
+                destination_folder = raw_input(
+                    "Please enter destination of the folder. (type 'exit' anytime to quit): ")
+                if destination_folder.strip() == 'exit':
+                    break
+                if os.path.exists(destination_folder) and os.path.isdir(destination_folder):
+                    create_sublime_project(destination_folder, config_path)
+                else:
+                    print("this does not seem to be a valid path :(")
+
+
 # OSX installation instructions
 def install_osx(app_path):
     install_path = '/opt/homebrew-cask/Caskroom/sublime-text3/build 3083/Sublime Text.app/Contents/SharedSupport/bin/subl'
+
     try:
         if not Helpers.is_installed(['brew', 'help', '&>/dev/null']):
             raise NotInstalledError(
                 'Error: Homebrew required to install Sublime Text.')
         print('Updating sources...')
+
         subprocess.call(['brew', 'update'], stdout=subprocess.PIPE)
         print('Installing Sublime Text...')
+
         os.environ['HOMEBREW_CASK_OPTS'] = '--appdir=/Applications'
         subprocess.call(['brew', 'tap', 'caskroom/versions'],
                         stdout=subprocess.PIPE, stderr=subprocess.PIPE)
         subprocess.call(['brew', 'cask', 'install', 'sublime-text3'],
                         stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+
         # symlink to 'sublime'
         subprocess.Popen(['ln', '-s', install_path, '/usr/local/bin/sublime'],
                          stdout=subprocess.PIPE, stderr=subprocess.PIPE)
@@ -107,7 +211,7 @@ def install_windows(app_path):
         if not Helpers.is_installed(['chocolatey']):
             raise NotInstalledError(
                 'Error: Chocolatey required to install Sublime Text.')
-        p = subprocess.Popen(['choco', 'install', '-y', 'sublimetext3'],
+        subprocess.Popen(['choco', 'install', '-y', 'sublimetext3'],
                              stdout=subprocess.PIPE, stderr=subprocess.PIPE)
         print('Installation complete...')
     except OSError as e:
@@ -117,73 +221,30 @@ def install_windows(app_path):
 
 # configuration instructions
 def config(config_path):
-    packages = config_path + '/Installed Packages'
-    settings = config_path + '/Packages/User'
-    # create the settings directories
-    Helpers.make_dir(config_path)
-    Helpers.make_dir(packages)
-    Helpers.make_dir(settings)
-    # install 'Package Control'
-    Helpers.install_package_control(config_path)
-    # copy the themes
-    Helpers.copytree('./themes', packages)
-    # copy the user preferences
-    Helpers.copytree('./user-settings', settings)
-    print('Configuration complete...')
+    configure = raw_input(
+        "Running now SublimeText configuration based on the templates provided\nThie will overwrite any existing configs if found, continue ? [y/n] : ")
 
+    if configure == "y":
+        packages = config_path + '/Installed Packages'
+        settings = config_path + '/Packages/User'
 
-# A class containting static helper methods used to perform the install +
-# config
-class Helpers():
-    # Check to see if an application is installed
-    @staticmethod
-    def is_installed(app_path):
-        try:
-            p = subprocess.Popen(
-                app_path, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-            os.kill(p.pid, signal.SIGTERM)
-            return True
-        except OSError as e:
-            return False
+        # create the settings directories
+        Helpers.make_dir(config_path)
+        Helpers.make_dir(packages)
+        Helpers.make_dir(settings)
 
-    # create a directory if it doesn't already exist
-    @staticmethod
-    def make_dir(dir):
-        try:
-            os.makedirs(dir)
-        # capture any non-file-creation errors
-        except OSError as exception:
-            if exception.errno != errno.EEXIST:
-                raise
+        # install 'Package Control'
+        Helpers.install_package_control(config_path)
 
-    # recursively copies all the files in a directory
-    #  required because shutil.copytree (sucks) can't overwrite directories
-    @staticmethod
-    def copytree(src, dst, symlinks=False, ignore=None):
-        for item in os.listdir(src):
-            s = os.path.join(src, item)
-            d = os.path.join(dst, item)
-            if os.path.isdir(s):
-                try:
-                    shutil.copytree(s, d, symlinks, ignore)
-                except OSError as exception:
-                    if exception.errno != errno.EEXIST:
-                        raise
-            else:
-                shutil.copy2(s, d)
+        # copy the themes
+        Helpers.copytree('./themes', packages)
 
-    @staticmethod
-    def install_package_control(config_path):
-        path = config_path + '/Installed Packages/Package Control.sublime-package'
-        url = 'ttps://packagecontrol.io/Package%20Control.sublime-package'
-        response = urllib2.urlopen(url)
-        CHUNK = 16 * 1024
-        with open(path, 'wb') as f:
-            while True:
-                chunk = response.read(CHUNK)
-                if not chunk:
-                    break
-                f.write(chunk)
+        # copy the user preferences
+        Helpers.copytree('./user-settings', settings)
+
+        print('SublimeText configuration completed successfully ...')
+    else:
+        print('SublimeText configuration not changed .. all is complete !')
 
 
 # Custom exception, indicates that a dependency is not installed
@@ -191,4 +252,8 @@ class NotInstalledError(Exception):
     pass
 
 if __name__ == '__main__':
-    main()
+    try:
+        main()
+    except KeyboardInterrupt:
+        print("\nCaught KeyboardInterrupt :(, terminating now ..")
+        sys.exit(1)
